Reset GlobalOpenTelemetry in tests
This commit ensures that GlobalOpenTelemetry is reset before and after each test in `TelemetryManagerTest.kt`. This prevents state leakage between tests, which can occur when `GlobalOpenTelemetry.set(sdk)` is called.

The reset is done by reflectively setting the `globalOpenTelemetry` field in the `GlobalOpenTelemetry` class to `null`. This is a workaround as `GlobalOpenTelemetry` does not provide a public reset method.

Signed-off-by: Meesum Ali <meesumdex@gmail.com>